Separated 'Projectile' and 'ProjectileVisualizer'
